remove legacy code references
